+++ apps/server/src/api/products/(products)/ledgerscan/schemas/transaction.schema.ts
@@ -1,8 +1,8 @@
-import mongoose from "mongoose"
+import { Schema } from "mongoose"
 
-export const TransactionSchema = new mongoose.Schema({
+export const TransactionSchema = new Schema({
   workspaceId: {
-    type: mongoose.Schema.Types.ObjectId,
+    type: Schema.Types.ObjectId,
     ref: "workspace",
     required: true
   },